Move all the host interfaces for the BBC MOS to a package.
The host interfaces have now moved into a package so that they can
be used more generically if people want. Specific routines that
interface with the host are separate from the base implementations
so that they don't have to be the way that you communicate with
the system.
